Fix hit testing
There were two issues. One was that the point of the sphere was just a
straight multiple of the direction vector. This failed for different
distances away from the camera.

The other was that the fudge factor for white space needed to be more
specific between vertical and horizontal white space since we had more
room vertically. Separated the two magic numbers out.